Instructor_Deployable_JW/requirements.txt
fastapi==0.100.1
uvicorn[standard]==0.23.2
motor==3.1.2
pymongo==4.3.3
requests==2.31.0
python-dotenv==1.0.0
pyyaml==6.0.1
pydantic==1.8.2
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.9
bcrypt==3.2.0
boto3
email-validator